추가 - 서버에 요청 수락/거절 로직 추가
