+++ KHFinalCertiHub/T3B1/src/main/java/com/kh/T3B1/study/controller/StudyController.java
@@ -62,7 +62,7 @@ public class StudyController {
 		
 		int result = studyService.insertStudy(study);
 		
-		if(result < 1) {
+		if(result == 0) {
 			session.setAttribute("errorMsg", "스터디 그룹 생성에 실패했습니다.");
 			return "redirect:/error";
 		}
@@ -472,7 +472,7 @@ public class StudyController {
 	@ResponseBody
 	@PostMapping(value="updateRecruit", produces="application/json; charset=UTF-8")
 	public String banMember(HttpSession session, String recruit, int studyNo, int boardNo) {
-		HashMap<String, String> resultObj = new HashMap<>(); // 결과값 반환용 객ㅔ
+		HashMap<String, String> resultObj = new HashMap<>(); // 결과값 반환용 객체
 		String result = "N"; // 실패 N 성공 Y
 		Member manager = (Member)session.getAttribute("loginMember");
