+++ KHFinalCertiHub/T3B1/src/main/resources/mappers/study-mapper.xml
@@ -282,13 +282,13 @@ INNER JOIN STUDY_MEMBER SM ON(M.MEMBER_NO = SM.MEMBER_NO) AND STUDY_NO = #{study
 		)
 	</insert>
 	
-	<insert id="insertStudyMember" parameterType="Study">
+	<insert id="insertStudyMember">
 		INSERT INTO STUDY_MEMBER(
 			MEMBER_NO,
 			STUDY_NO,
 			CONFIRM_DATE
 		) VALUES (
-			#{managerNo},
+			#{memberNo},
 			STUDY_SEQ.CURRVAL,
 			SYSDATE
 		)
@@ -321,6 +321,6 @@ INNER JOIN STUDY_MEMBER SM ON(M.MEMBER_NO = SM.MEMBER_NO) AND STUDY_NO = #{study
   	   SET STUDY_RECRUIT = #{recruit}
   	 WHERE STUDY_NO = #{studyNo}
   </update>
-
+	
 </mapper>
